Separate common functionality of SolanaLedgerBridgeKeyring into a parent class LedgerBridgeKeyring that EthLedgerBridgeKeyring will inherit from
