+++ components/brave_wallet_ui/common/hardware/ledgerjs/sol_ledger_bridge_keyring.ts
@@ -3,21 +3,14 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import { LEDGER_HARDWARE_VENDOR } from 'gen/brave/components/brave_wallet/common/brave_wallet.mojom.m.js'
 import { BraveWallet } from '../../../constants/types'
-import { getLocale } from '../../../../common/locale'
 import { LedgerSolanaKeyring } from '../interfaces'
-import { HardwareVendor } from '../../api/hardware_keyrings'
 import {
   GetAccountsHardwareOperationResult,
-  HardwareOperationResult,
   SignHardwareOperationResult
 } from '../types'
 import {
-  LEDGER_BRIDGE_URL,
   LedgerCommand,
-  UnlockResponse,
-  LedgerFrameCommand,
   GetAccountResponse,
   GetAccountResponsePayload,
   SignTransactionResponse,
@@ -25,44 +18,16 @@ import {
   LedgerErrorsCodes,
   LedgerError
 } from './ledger-messages'
-import { LedgerTrustedMessagingTransport } from './ledger-trusted-transport'
 import { hardwareDeviceIdFromAddress } from '../hardwareDeviceIdFromAddress'
+import LedgerBridgeKeyring from './ledger_bridge_keyring'
 
-// storybook compiler thinks `randomUUID` doesn't exist
-const randomUUID = () => (window.crypto as Crypto & { randomUUID: () => string }).randomUUID()
-export default class SolanaLedgerBridgeKeyring implements LedgerSolanaKeyring {
-  private deviceId: string
-  private onAuthorized?: () => void
-  private transport?: LedgerTrustedMessagingTransport
-  private bridge?: HTMLIFrameElement
-  private readonly frameId: string
-
+export default class SolanaLedgerBridgeKeyring extends LedgerBridgeKeyring implements LedgerSolanaKeyring {
   constructor (onAuthorized?: () => void) {
-    this.onAuthorized = onAuthorized
-    this.frameId = randomUUID()
+    super(onAuthorized)
   }
 
   coin = (): BraveWallet.CoinType => {
     return BraveWallet.CoinType.SOL
-  }
-
-  type = (): HardwareVendor => {
-    return LEDGER_HARDWARE_VENDOR
-  }
-
-  unlock = async (): Promise<HardwareOperationResult> => {
-    const data = await this.sendCommand<UnlockResponse>({
-      id: LedgerCommand.Unlock,
-      origin: window.origin,
-      command: LedgerCommand.Unlock
-    })
-
-    if (data === LedgerErrorsCodes.BridgeNotReady ||
-        data === LedgerErrorsCodes.CommandInProgress) {
-      return this.createErrorFromCode(data)
-    }
-
-    return data.payload
   }
 
   getAccounts = async (from: number, to: number): Promise<GetAccountsHardwareOperationResult> => {
@@ -112,21 +77,6 @@ export default class SolanaLedgerBridgeKeyring implements LedgerSolanaKeyring {
     return { success: true, payload: responsePayload.signature }
   }
 
-  private readonly createBridge = (targetUrl: string) => {
-    return new Promise<HTMLIFrameElement>((resolve) => {
-      const element = document.createElement('iframe')
-      element.id = this.frameId
-      element.src = (new URL(targetUrl)).origin + `?targetUrl=${encodeURIComponent(window.origin)}`
-      element.style.display = 'none'
-      element.allow = 'hid'
-      element.onload = () => {
-        this.bridge = element
-        resolve(element)
-      }
-      document.body.appendChild(element)
-    })
-  }
-
   private readonly getAccountsFromDevice = async (paths: string[], skipZeroPath: boolean): Promise<GetAccountsHardwareOperationResult> => {
     let accounts = []
     const zeroPath = this.getPathForIndex(0)
@@ -172,39 +122,7 @@ export default class SolanaLedgerBridgeKeyring implements LedgerSolanaKeyring {
     return { success: true, payload: accounts }
   }
 
-  sendCommand = async <T> (command: LedgerFrameCommand): Promise<T | LedgerErrorsCodes > => {
-    if (!this.bridge && !this.hasBridgeCreated()) {
-      this.bridge = await this.createBridge(LEDGER_BRIDGE_URL)
-    }
-    if (!this.bridge || !this.bridge.contentWindow) {
-      return LedgerErrorsCodes.BridgeNotReady
-    }
-    if (!this.transport) {
-      this.transport = new LedgerTrustedMessagingTransport(
-        this.bridge.contentWindow,
-        LEDGER_BRIDGE_URL,
-        this.onAuthorized
-      )
-    }
-    return this.transport.sendCommand(command)
-  }
-
-  cancelOperation = async () => { }
-
-  private readonly createErrorFromCode = (code: LedgerErrorsCodes): HardwareOperationResult => {
-    switch (code) {
-      case LedgerErrorsCodes.BridgeNotReady:
-        return { success: false, error: getLocale('braveWalletBridgeNotReady'), code: code }
-      case LedgerErrorsCodes.CommandInProgress:
-        return { success: false, error: getLocale('braveWalletBridgeCommandInProgress'), code: code }
-    }
-  }
-
   private readonly getPathForIndex = (index: number): string => {
     return `44'/501'/${index}'/0'`
-  }
-
-  private readonly hasBridgeCreated = (): boolean => {
-    return document.getElementById(this.frameId) !== null
   }
 }
